Use cutomhook and common code for counter
