+++ 08-custom-hook/src/components/BackwardCounter.js
@@ -1,19 +1,10 @@
-import { useState, useEffect } from 'react';
+import useCounter from '../hooks/use-counter';
 
 import Card from './Card';
 
 const BackwardCounter = () => {
-  const [counter, setCounter] = useState(0);
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setCounter((prevCounter) => prevCounter - 1);
-    }, 1000);
-
-    return () => clearInterval(interval);
-  }, []);
-
-  return <Card>{counter}</Card>;
+    const counter = useCounter(false);
+    return <Card>{counter}</Card>;
 };
 
 export default BackwardCounter;
