+++ 08-custom-hook/src/components/ForwardCounter.js
@@ -1,18 +1,8 @@
-import { useState, useEffect } from 'react';
-
+import useCounter from '../hooks/use-counter';
 import Card from './Card';
 
 const ForwardCounter = () => {
-  const [counter, setCounter] = useState(0);
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setCounter((prevCounter) => prevCounter + 1);
-    }, 1000);
-
-    return () => clearInterval(interval);
-  }, []);
-
+  const counter = useCounter();
   return <Card>{counter}</Card>;
 };
